+++ packages/components/src/components/input-password/input-password.tsx
@@ -18,6 +18,7 @@ import {
   getPrefixedTagNames,
   hasPropValueChanged,
   implicitSubmit,
+  syncFormState,
   validateProps,
 } from '../../utils';
 import { InputBase } from '../common/input-base/input-base';
@@ -149,7 +150,6 @@ export class InputPassword {
     if (this.inputElement && this.inputElement.value !== this.parsedValue) {
       this.inputElement.value = this.parsedValue;
     }
-    this.internals?.setFormValue(this.parsedValue);
   }
 
   public connectedCallback(): void {
@@ -178,18 +178,12 @@ export class InputPassword {
     return hasPropValueChanged(newVal, oldVal);
   }
 
-  public componentDidLoad(): void {
-    this.internals?.setFormValue(this.parsedValue);
-  }
-
   public componentDidRender(): void {
-    if (!this.disabled && !this.readOnly) {
-      this.internals?.setValidity(
-        this.inputElement.validity,
-        this.inputElement.validationMessage || ' ',
-        this.inputElement
-      );
-    }
+    syncFormState(this.internals, this.inputElement, {
+      disabled: this.disabled,
+      readOnly: this.readOnly,
+      value: this.parsedValue,
+    });
   }
 
   public componentWillUpdate(): void {